+++ src/Sesion/Home.js
@@ -1,0 +1,12 @@
+import { onNavigate } from "../main.js";
+
+ export const Home = (rootElement)=>{
+   const button2 = ` <button class="logo" id="LogoButton">Asia Mania</button>  `  
+ rootElement.innerHTML= button2;
+ 
+ const IniciarSecion = document.getElementById('LogoButton'); 
+ IniciarSecion.addEventListener('click', function(){
+    onNavigate('/Login');
+ }
+ ); 
+ }
