+++ src/css/HomeStyle.css
@@ -1,0 +1,10 @@
+/* Cuando la pantalla sea más chica que 480px aplica estos estilos */
+@media (max-width: 480px) {
+.logo {
+    background-color: blue;
+}
+}
+/* Pero si mi pantalla es mayor a 480 px aplica estos estilos siempre y cuando el tamaño sea menor a 768 px */
+@media (max-width: 768px) {
+
+}
