+++ src/index.html
@@ -41,7 +41,9 @@
 <html lang="en">
 <head>
   <meta charset="UTF-8">
-  <title>Prueba de autenticacion firebase para google (sdk de Javascript)</title>
+  <link rel="stylesheet" href="./css/HomeStyle.css" />
+
+  <title>Asia Mania</title>
   <style>
     #datosuser{
       font-size: .8em;
